Moved argument parsing out of int main
I'm definitely in the "int main should be nothing more than a series of
function calls" camp, and it was bothering me that it was doing all that
argument parsing. So I moved the parsing into a new function.

I also fixed a glitch in the argument parsing caused by my own
misunderstanding of how getopt works; I didn't realize it automatically
reported specific errors, and I didn't expect it to run my switch
statement with a `-?` argument. So it was reporting two errors where it
now only reports the one.