Removed all autoconf files; added genconf
In particular, I dislike having visible Makefile.am and configure.ac
files all over the place, all having nearly the same name as generated
files that actually serve a purpose (i.e. Makefile.in and configure).

So enough of that! They are now rolled into a single script that, as a
bonus, also runs all the appropriate autoconf commands to regenerate the
files that I actually want.

Except it doesn't do that yet. At the moment, all it does is demonstrate
its ability to create the needed files. Next up is actually handling
them.